@@ -10,7 +10,7 @@ Gem::Specification.new do |s|
 
   s.required_ruby_version = ">= 2.2.2"
 
-  s.add_runtime_dependency('rails', '>= 4.2.6')
+  s.add_runtime_dependency('rails', '>= 5.0.0')
   s.add_runtime_dependency('shopify_api', '>= 4.3.2')
   s.add_runtime_dependency('omniauth-shopify-oauth2', '~> 1.1.11')
 
