@@ -8,39 +8,39 @@ class LocalizationController < ActionController::Base
   before_action :set_locale
 
   def index
-    render text: I18n.locale
+    head :ok
   end
 end
 
 class LocalizationTest < ActionController::TestCase
   tests LocalizationController
 
-  test "falls back to I18n.default if locale param is not present" do
+  setup do
     I18n.available_locales = [:en, :de, :es, :ja, :fr]
+  end
+
+  test "falls back to I18n.default if locale param is not present" do
     I18n.default_locale = :ja
 
     with_test_routes do
       get :index
-      assert 'ja', response.body
+      assert_equal :ja, I18n.locale
     end
   end
 
   test "set I18n.locale to passed locale param" do
-    I18n.available_locales = [:en, :de, :es, :ja, :fr]
-
     with_test_routes do
-      get :index, locale: 'de'
-      assert 'de', response.body
+      get :index, params: { locale: 'de' }
+      assert_equal :de, I18n.locale
     end
   end
 
   test "falls back to I18n.default if locale is not supported" do
-    I18n.available_locales = [:en, :de, :es, :ja, :fr]
     I18n.default_locale = :en
 
     with_test_routes do
-      get :index, locale: 'cu'
-      assert 'en', response.body
+      get :index, params: { locale: 'invalid_locale' }
+      assert_equal :en, I18n.locale
     end
   end
 
