@@ -3,12 +3,7 @@ module ShopifyApp
     extend ActiveSupport::Concern
 
     included do
-      if Rails.version >= '5.0'
-        skip_before_action :verify_authenticity_token, raise: false
-      else
-        skip_before_action :verify_authenticity_token
-      end
-
+      skip_before_action :verify_authenticity_token, raise: false
       before_action :verify_request
     end
 
