@@ -35,17 +35,29 @@ class AppProxyVerificationTest < ActionController::TestCase
     assert query_string_valid? 'shop=some-random-store.myshopify.com&path_prefix=%2Fapps%2Fmy-app&timestamp=1466106083&foo=bar&baz[1]&baz[2]=b&baz[c[0]]=whatup&baz[c[1]]=notmuch&signature=bbf3aa60e098f08919a2ea4c64a388414f164e6a117a63b03479ac7aa9464b4f'
   end
 
-  test 'request_should_fail' do
+  test 'request with invalid signature should fail' do
     with_test_routes do
-      get(:basic, shop: 'some-random-store.myshopify.com', path_prefix: '/apps/my-app', timestamp: '1466106083', signature: 'wrong233558b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd')
-      assert_response(:unauthorized)
+      invalid_params = {
+        shop: 'some-random-store.myshopify.com',
+        path_prefix: '/apps/my-app',
+        timestamp: '1466106083',
+        signature: 'wrong233558b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd'
+      }
+      get :basic, params: invalid_params
+      assert_response :unauthorized
     end
   end
 
-  test 'request_should_pass' do
+  test 'request with a valid signature should pass' do
     with_test_routes do
-      get(:basic, shop: 'some-random-store.myshopify.com', path_prefix: '/apps/my-app', timestamp: '1466106083', signature: 'f5cd7233558b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd')
-      assert_response(:ok)
+      valid_params = {
+        shop: 'some-random-store.myshopify.com',
+        path_prefix: '/apps/my-app',
+        timestamp: '1466106083',
+        signature: 'f5cd7233558b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd'
+      }
+      get :basic, params: valid_params
+      assert_response :ok
     end
   end
 
